+++ src/main/java/org/dinosaur/foodbowl/domain/member/repository/MemberRoleRepository.java
@@ -1,4 +1,4 @@
-package org.dinosaur.foodbowl.domain.member.dao;
+package org.dinosaur.foodbowl.domain.member.repository;
 
 import org.dinosaur.foodbowl.domain.member.entity.MemberRole;
 import org.springframework.data.repository.Repository;
